Added pdfinterleave.bash script
The pdfinterleave.bash script interleaves pages from two distinct PDF
files and produces an output PDF file. The odd pages are taken from a
first PDF file and the even pages are taken from a second PDF file passed
respectively as first and second argument. Note that the even pages will
be used in the reverse order (last one first)

This is useful if you have two-sided documents scanned each side on a
different file as it can happen when using a one-sided Automatic Document
Feeder (ADF).

It does a similar job to :
https://github.com/weltonrodrigo/pdfapi2/blob/46434ab3f108902db2bc49bcf06f66544688f553/contrib/pdf-interleave.pl
but only requires bash (> 4.0) and poppler utils.